fix(common): move kms account to its own export
