+++ packages/common/src/account/kms/getPublicKey.ts
