+++ packages/common/src/exports/kms.ts
@@ -1,0 +1,1 @@
+export { createKmsAccount, type CreateKmsAccountOptions, type KMSAccount } from "../account/kms/createKmsAccount";
